fix: update of pixelated icons and logos
A large part of the .png images were in pixelated quality at higher resolutions on the website 'herbsjs.org'. With this fix, the images and icons, in addition to the ones that were mentioned in the issue, such as the discord, twitter and stack overflow icons were also fixed and transformed into .svg.